consolidated tests
